+++ src/components/Header/Header.tsx
@@ -1,7 +1,6 @@
 import { faSun, faMoon } from "@fortawesome/free-regular-svg-icons";
 import { faVolumeUp, faVolumeMute } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { Howler } from "howler";
 import localForage from "localforage";
 import React, { useState } from "react";
 import {
