+++ src/models/ApplicationSettings.ts
