--- src/models/ConnectionStatus.ts
@@ -1,5 +1,0 @@
-export enum ConnectionStatus {
-  SINGLE_PLAYER,
-  IS_HOST,
-  IS_CLIENT,
-}
