+++ src/models/GameSettingsType.ts
