+++ src/store/application/applicationConstants.ts
@@ -1,5 +1,3 @@
-export const APPLICATION_SET_PLAYER_NAME = "application/setPlayerName";
-
 export const APPLICATION_SET_VOLUME = "application/setVolume";
 export const APPLICATION_SET_IS_DARK = "application/setIsDark";
 export const APPLICATION_SET_LOADED_SETTINGS = "application/setLoadedSettings";
