--- src/store/middleware/middlewareConstants.ts
@@ -1,1 +1,0 @@
-export const PLAY_AUDIO_ACTION = "application/playAudio";
